requirements.txt
aioredis==1.3.1
altgraph==0.17.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
distlib==0.3.6
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-pyproject==1.0.6
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
filelock==3.8.0
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.2
onnx==1.12.0
packaging==21.3
pandas==1.4.3
pefile==2022.5.30
phonenumbers==8.12.55
Pillow==9.2.0
pip-box==0.4.5
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyotp==2.7.0
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.18.1
python-dateutil==2.8.2
python-gettext==4.1
pytz==2022.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
swagger-spec-validator==2.7.6
tmdbv3api==1.7.7
toml==0.10.2
twilio==7.14.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.2
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
waitress==2.1.2
whitenoise==6.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
zope.interface==5.0.0